Settings split Tier 2: SettingsApp domain (Domain 4 of 4)
Move app-level properties out of Settings into a new SettingsApp sub-object
(updates channel, daily backup hour, dev/platform flags, water level/refill,
profile favorites/selected/hidden/current, deviceId, pocketPairingToken,
hasQuick3D, use12HourTime, isDebugBuild). Settings.h shrinks 466 → 298 lines.
