+++ qml/components/layout/LayoutItemDelegate.qml
@@ -48,7 +48,7 @@ Item {
                 action: "togglePreset:espresso",
                 longPressAction: "navigate:profiles",
                 doubleclickAction: "navigate:profiles",
-                backgroundColor: Settings.selectedFavoriteProfile === -1 ? Theme.highlightColor : Theme.primaryColor
+                backgroundColor: Settings.app.selectedFavoriteProfile === -1 ? Theme.highlightColor : Theme.primaryColor
             }
             case "steam": return {
                 emoji: "qrc:/icons/steam.svg",
