+++ qml/components/layout/items/CustomItem.qml
@@ -217,7 +217,7 @@ Item {
                 statusImg((scaleOn && !flowScale) ? statusConnected : statusDisconnected))
         // Time
         var now = new Date()
-        result = result.replace(/%TIME%/g, Qt.formatTime(now, Settings.use12HourTime ? "h:mmap" : "hh:mm"))
+        result = result.replace(/%TIME%/g, Qt.formatTime(now, Settings.app.use12HourTime ? "h:mmap" : "hh:mm"))
         result = result.replace(/%DATE%/g, Qt.formatDate(now, "yyyy-MM-dd"))
         // Convert any emoji Unicode in the result to <img> tags to avoid
         // CoreText/ImageIO crash from Apple Color Emoji PNG decoding on render thread
