+++ qml/pages/VisualizerMultiImportPage.qml
@@ -662,7 +662,7 @@ Page {
                             Text { text: TranslationManager.translate("visualizerImport.shotTime", "Shot time:"); color: Theme.textSecondaryColor; font: Theme.captionFont }
                             Text {
                                 text: selectedShot && selectedShot.start_time ?
-                                      new Date(selectedShot.start_time).toLocaleString(Qt.locale(), Settings.use12HourTime ? "MMM d, yyyy h:mm AP" : "MMM d, yyyy HH:mm") : "-"
+                                      new Date(selectedShot.start_time).toLocaleString(Qt.locale(), Settings.app.use12HourTime ? "MMM d, yyyy h:mm AP" : "MMM d, yyyy HH:mm") : "-"
                                 color: Theme.textColor
                                 font: Theme.captionFont
                             }
